have Session implement the Logger interface
This is a step toward logging the session PID in each Session
log message, which will make it easier to debug multiseat setups.

The PID is set to 0 when the Session process terminates, so it is not
possible to follow a Session's complete lifetime in the logs.  But
it's better than nothing.

A future commit could easily change this to log some other identifier,
such as the associated Display or Seat name, or the PID before the
process terminated.