Enable the C preprocessor when loading X resources
A common use for X resource files is customization of a user's terminal. Widely
used color schemes use the C preprocessor to avoid repetition. See for example
solarized[1] or base-16[2].
  
The counter argument, that -nocpp leads to a performance improvement for the
general audience, was disproved by igel and roelof000 in lp:1244578. To
summarize, the speed improvement is in the range of 0.01 seconds and the
commands will not even be executed if you do not have X resource files.
  
[1] https://github.com/altercation/solarized/blob/master/xresources/solarized
[2] https://github.com/chriskempson/base16-xresources