Terminate leftover processes in greeter session
It can happen that the greeter session does not properly clean up itself on
logout. This causes leaked processes like settings-daemon which act on the user
session as they share the same $DISPLAY. Ask logind to terminate remaining
processes in the greeter session on closing, which is more robust than trying
to fix every greeter.