Explicitly set file mode when creating guest home directory
The fix for CVE-2017-7358 (bzr commit 2478) adjusted the guest home
directory creation to use mkdir in some instances. In those instances,
the home directory is world readable for a short amount of time before
tmpfs is mounted on top of the directory with a 700 file mode. There is
no information leak here since data isn't written to the directory
before the tmpfs mount occurs but it is best to ask mkdir to set the
file mode to 700 from the start.

Thanks to Alexander Peslyak (Solar Designer) for catching this mistake.