WSC: set the GIO encoding for the from USC pipe to binary
We always read raw data from the pipe, thus this should have been set
since this file was written. However, for some reason this has nerver
been set, and for some unknown reason this has never been a problem on
Xenial port of UT. However, not having this caused an infinite hang on
g_io_channel_read_chars(), presumebly because it's trying to interpret
a binary data as an incomplete UTF-8 character.

From:
https://gitlab.com/ubports/core/packaging/lightdm/-/commit/6616c8a8b6451751d130696f8edcf8aff0c11c35
Author: Ratchanan Srirattanamet