liblightdm-gobject/language: check before calling get_locale_name
On libc's that doesn't have multiple locale, for example musl, calling
get_locale_name would result in build error with message
LC_IDENTIFICATION undeclared. This bug seems to affect BSD systems too.

I found a very old thread about this error here [1], but I hit this bug
when porting stuff for musl on Gentoo Linux

"LC_IDENTIFICATION - this is not a user-visible category but contains
information about the locale itself and is rarely useful for users or
developers (but is listed here for completeness sake)."

Signed-off-by: listout <listout@protonmail.com>