tests: New FENCE script command to ensure event order
By default, emitted status text is allowed to match a status matcher
line from anywhere in the script.  Thus, a script like the
following:

    #?*FOO
    #?BAR

will happily accept a sequence of events like this:

    BAR
    *FOO

This loose ordering avoids test flakiness in the presence of
concurrent events, but can be problematic if the test wants to assert
that the BAR event is an effect of the FOO command and not
coincidental.  (A concrete example: assert that a greeter was
launched *because* the user session terminated, not in parallel with
the briefly-lived user session.)

Script authors may be tempted to assert causality by introducing a
WAIT like this:

    #?*WAIT
    #?*FOO
    #?BAR

but that would not have the desired effect because it will still
accept a BAR event before the FOO command (or even the WAIT command)
is executed.

The new FENCE command helps script authors assert causality by
ensuring that an emitted status does not match a line on the other
side of the fence.  For example, the following script works as
expected and asserts a causal relationship between FOO and BAR:

    #?*WAIT
    #?*FENCE
    #?*FOO
    #?BAR

The above script only works if the WAIT is long enough to sufficiently
rule out coincidence.  To see why, note that the following sequence of
events would be accepted by the above script:

    *WAIT
    *FENCE
    BAR
    *FOO

The order of the FENCE and FOO should not be switched, otherwise it
introduces a theoretical race condition: If the BAR event is caused by
the FOO command, and the BAR event happens to arrive before the FENCE
is executed, then the test will fail when it should not.  (Technically
it is not possible to lose the race right now because of how
run_commands is implemented, but scripts should not rely on an
implementation detail.  Besides, it doesn't eliminate the need for the
WAIT.)

This new command will be used in a future commit to add a regression
test for a Wayland multiseat bug.