tests: Use unix:dir instead of unix:tmpdir for dbus-daemon socket
Starting with dbus-daemon v1.15.2, unix:tmpdir no longer causes
dbus-daemon to create abstract sockets (for security reasons), so
unix:tmpdir is now equivalent to unix:dir.  (Also, some systems don't
support abstract sockets.)  Switch to unix:dir so that it is clear
that the tests expect a socket file to be created in the filesystem.