Add support for changing the home directory in PAM modules
PAM modules such as pam_mklocaluser may change or even create the home
directory. Currently, LightDM assumes that the home directory will not change
when opening the PAM session, the user's home directory is obtained via
getpwent() after authentication but before opening the session.  Fix this by
trying to update the user's home directory from the HOME environment variable
from PAM after opening the session.
Furthermore, if the Xauthority file is not stored in a system directory, the
daemon hardcodes its path to the user's home directory and passes it as an
absolute path to the session child.  Fix this by passing it as a relative path
so that the actual path can be constructed after the PAM session has been
opened and the home directory has potentially been updated.